src/tools/web_fetch.py
"""
工具接口 - 网页抓取
"""
from typing import Optional, Dict, Any
import requests
from bs4 import BeautifulSoup


class WebFetchTool:
    """网页抓取工具"""
    
    def __init__(self, timeout: int = 30):
        self.timeout = timeout
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        })
    
    def fetch(self, url: str) -> Dict[str, Any]:
        """
        抓取网页内容
        
        Args:
            url: 目标URL
            
        Returns:
            {
                "status": "success" | "error",
                "url": url,
                "html": html内容,
                "text": 纯文本内容,
                "title": 页面标题,
                "error": 错误信息（如果有）
            }
        """
        try:
            response = self.session.get(url, timeout=self.timeout)
            response.raise_for_status()
            
            soup = BeautifulSoup(response.text, 'html.parser')
            
            # 移除script和style标签
            for script in soup(["script", "style"]):
                script.decompose()
            
            return {
                "status": "success",
                "url": url,
                "html": response.text,
                "text": soup.get_text(separator='\n', strip=True),
                "title": soup.title.string if soup.title else "",
            }
        except Exception as e:
            return {
                "status": "error",
                "url": url,
                "html": "",
                "text": "",
                "title": "",
                "error": str(e)
            }
    
    def fetch_youtube_channel(self, channel_url: str) -> Dict[str, Any]:
        """
        抓取YouTube频道页面
        
        TODO: 实现YouTube频道数据解析
        - 频道名称、订阅数
        - 视频列表
        - About页面信息
        """
        result = self.fetch(channel_url)
        
        if result["status"] == "success":
            # TODO: 解析YouTube特定数据结构
            pass
        
        return result
